[ADD] Litepost component 분리
