--- src/app/(providers)/(root)/(public)/survey-lite/_components/Button.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-interface ButtonProps {
-  onClick: () => void;
-  children: React.ReactNode;
-}
-
-const Button: React.FC<ButtonProps> = ({onClick, children}) => {
-  return <button onClick={onClick}>{children}</button>;
-};
-
-export default Button;
