+++ src/app/(providers)/(root)/(public)/survey-lite/page.tsx
@@ -1,21 +1,19 @@
 'use client';
 
-import {deleteliteSurveyPostById, getLiteSurveyPosts} from '@/app/api/litepagefirbaseApi';
+import {deleteliteSurveyPostById, getLiteSurveyPosts, updateLikesCount} from '@/app/api/litepagefirbaseApi';
 import {litePost} from '@/app/api/typePost';
 import {auth, db} from '@/firebase';
 import {Button} from '@nextui-org/react';
 import {useQuery} from '@tanstack/react-query';
 import {collection, deleteDoc, doc, getDoc, getDocs, setDoc, updateDoc} from 'firebase/firestore';
 import {useEffect, useState} from 'react';
-import {FaHeart, FaRegHeart} from 'react-icons/fa';
-import {FaRegCircleUser} from 'react-icons/fa6';
-import {GrView} from 'react-icons/gr';
 import {LuPencilLine} from 'react-icons/lu';
 import Swal from 'sweetalert2';
 import Banner from '../../(main)/_components/carousel/Banner';
 import LiteSurveyCreateModal from '../../(main)/_components/modal/CreateModal';
 import LiteSurveyModal from '../../(main)/_components/modal/SurveyModal';
 import UpdateModal from '../../(main)/_components/modal/UpdateModal';
+import LitePostComponent from './_components/LitePostComponent';
 
 export default function SurveyLitePage() {
   const [selectedPost, setSelectedPost] = useState<litePost | null>(null);
@@ -154,14 +152,6 @@ export default function SurveyLitePage() {
     }
   };
 
-  // 새로운 게시물 알려주기
-  const isWithin24Hours = (createdAt: Date): boolean => {
-    const currentTime = new Date();
-    const timeDifference = currentTime.getTime() - createdAt.getTime();
-    const hoursDifference = timeDifference / (1000 * 60 * 60);
-    return hoursDifference <= 24;
-  };
-
   // 좋아하는 게시물을 서브컬렉션으로 저장하기
   const updateLikedPostsSubcollection = async (userId: string, postId: string, isLiked: boolean) => {
     try {
@@ -186,27 +176,17 @@ export default function SurveyLitePage() {
     }
     if (userId) {
       try {
-        const postRef = doc(db, 'litesurveyposts', postId);
-        const postSnapshot = await getDoc(postRef);
+        // 좋아요 수 카운트 함수
+        await updateLikesCount(postId, userId, likedPosts);
 
-        if (postSnapshot.exists()) {
-          const currentLikes = postSnapshot.data().likes || 0;
-          const updatedLikes = likedPosts[postId] ? currentLikes - 1 : currentLikes + 1;
+        // 사용자 문서 업데이트: 좋아하는 게시물의 ID를 업데이트합니다.
+        await updateLikedPostsSubcollection(userId, postId, !likedPosts[postId]);
 
-          // 좋아요 수 업데이트
-          await updateDoc(postRef, {likes: updatedLikes});
-
-          // 사용자 문서 업데이트: 좋아하는 게시물의 ID를 업데이트합니다.
-          await updateLikedPostsSubcollection(userId, postId, !likedPosts[postId]);
-
-          // likedPosts 상태 업데이트
-          setLikedPosts(prevState => ({
-            ...prevState,
-            [postId]: !prevState[postId],
-          }));
-        } else {
-          console.error(`게시물 ID ${postId}에 해당하는 문서가 존재하지 않습니다.`);
-        }
+        // likedPosts 상태 업데이트
+        setLikedPosts(prevState => ({
+          ...prevState,
+          [postId]: !prevState[postId],
+        }));
       } catch (error) {
         console.error('좋아요 수 업데이트 중 오류:', error);
       }
@@ -255,86 +235,17 @@ export default function SurveyLitePage() {
               {liteSurveyData && liteSurveyData.length > 0 ? (
                 <div className="post-container grid grid-cols-4 gap-4">
                   {liteSurveyData?.sort(sortByCreatedAt).map(litepost => (
-                    <div key={litepost.id}>
-                      <div className="h-[13.4375rem] bg-white border-1 border-[#C1C5CC] flex-col justify-between rounded-md p-4">
-                        <div className="top-content h-[5.625rem]">
-                          <div className="flex justify-between items-center mb-4">
-                            <div className="flex gap-2">
-                              {/* <p className="bg-[#0051FF] text-[#D6FF00] w-14 p-1 text-center rounded-full font-semibold text-xs">
-                                Lite
-                              </p> */}
-                              <p
-                                className={`bg-[#D6FF00] text-black w-14 p-1 text-center rounded-full font-semibold text-xs ${
-                                  isWithin24Hours(litepost.createdAt) ? '' : 'hidden'
-                                }`}
-                              >
-                                {isWithin24Hours(litepost.createdAt) ? 'New🔥' : ''}
-                              </p>
-                              <button
-                                className="toggle-menu w-8 h-7"
-                                onClick={() => onClickUpdateDeleteMenuToggle(litepost.id)}
-                              >
-                                {userId === litepost.userId && (menuStates[litepost.id] ? 'X' : '⁝')}
-                              </button>
-                              {menuStates[litepost.id] && (
-                                <div className="gap-2">
-                                  <button
-                                    className="w-8 h-7 text-blue-800 hover:bg-gray-100"
-                                    onClick={() => onClickUpdateButton(litepost.id)}
-                                  >
-                                    수정
-                                  </button>
-                                  <button
-                                    className="w-8 h-7 text-red-500 hover:bg-gray-100"
-                                    onClick={() => onClickDeleteButton(litepost.id)}
-                                  >
-                                    삭제
-                                  </button>
-                                </div>
-                              )}
-                            </div>
-                            <button
-                              onClick={() => onClickLikedPostHandler(litepost.id)}
-                              className="like-button w-12 h-[1.25rem] flex justify-evenly items-center text-[#0051FF]"
-                            >
-                              {litepost.likes} {likedPosts[litepost.id] ? <FaHeart /> : <FaRegHeart />}
-                            </button>
-                          </div>
-                          <a onClick={() => onClickPosthandler(litepost)} className="cursor-pointer">
-                            <div className="flex justify-between">
-                              <div>
-                                <p className="text-xs text-[#666] mb-4">
-                                  작성일 |{' '}
-                                  {litepost.createdAt
-                                    ? litepost.createdAt.toLocaleString('ko-KR', {
-                                        year: 'numeric',
-                                        month: '2-digit',
-                                        day: '2-digit',
-                                      })
-                                    : '2099.12.31'}
-                                </p>
-                              </div>
-                              {/* <p className="text-xs text-[#666] mb-4">
-                                마감일 | {litepost.deadlineDate ? litepost.deadlineDate.toLocaleString() : '2099.12.31'}
-                              </p> */}
-                            </div>
-                            <h3 className="text-lg font-bold">{litepost.title}</h3>
-                          </a>
-                        </div>
-                        <div className="bottom-content flex items-end">
-                          <div className="flex justify-between items-center mt-[3.125rem] w-full border-t-1 ">
-                            <div className="user flex mt-4 gap-2">
-                              <FaRegCircleUser />
-                              <p className="font-semibold">{litepost.nickname}</p>
-                            </div>
-                            <div className="viewer flex mt-4 gap-2 text-[#818490]">
-                              <GrView />
-                              {litepost.views}
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-                    </div>
+                    <LitePostComponent
+                      key={litepost.id}
+                      litepost={litepost}
+                      onClickPosthandler={onClickPosthandler}
+                      onClickUpdateDeleteMenuToggle={onClickUpdateDeleteMenuToggle}
+                      onClickUpdateButton={onClickUpdateButton}
+                      onClickDeleteButton={onClickDeleteButton}
+                      onClickLikedPostHandler={onClickLikedPostHandler}
+                      likedPosts={likedPosts}
+                      menuStates={menuStates}
+                    />
                   ))}
                 </div>
               ) : (
